+++ libcloud/compute/drivers/gce.py
@@ -64,7 +64,28 @@ class GCEResponse(GoogleResponse):
 
 
 class GCEConnection(GoogleBaseConnection):
-    """Connection class for the GCE driver."""
+    """
+    Connection class for the GCE driver.
+
+    GCEConnection extends :class:`google.GoogleBaseConnection` for 2 reasons:
+      1. modify request_path for GCE URI.
+      2. Implement gce_params functionality described below.
+
+    If the parameter gce_params is set to a dict prior to calling request(),
+    the URL parameters will be updated to include those key/values FOR A
+    SINGLE REQUEST. If the response contains a nextPageToken,
+    gce_params['pageToken'] will be set to its value. This can be used to
+    implement paging in list:
+
+    >>> params, more_results = {'maxResults': 2}, True
+    >>> while more_results:
+    ...     driver.connection.gce_params=params
+    ...     driver.ex_list_urlmaps()
+    ...     more_results = 'pageToken' in params
+    ...
+    [<GCEUrlMap id="..." name="cli-map">, <GCEUrlMap id="..." name="lc-map">]
+    [<GCEUrlMap id="..." name="web-map">]
+    """
     host = 'www.googleapis.com'
     responseCls = GCEResponse
 
@@ -76,6 +97,138 @@ class GCEConnection(GoogleBaseConnection):
                                             **kwargs)
         self.request_path = '/compute/%s/projects/%s' % (API_VERSION,
                                                          project)
+        self.gce_params = None
+
+    def pre_connect_hook(self, params, headers):
+        """
+        Update URL parameters with values from self.gce_params.
+
+        @inherits: :class:`GoogleBaseConnection.pre_connect_hook`
+        """
+        params, headers = super(GCEConnection, self).pre_connect_hook(params,
+                                                                      headers)
+        if self.gce_params:
+            params.update(self.gce_params)
+        return params, headers
+
+    def request(self, *args, **kwargs):
+        """
+        Perform request then do GCE-specific processing of URL params.
+
+        @inherits: :class:`GoogleBaseConnection.request`
+        """
+        response = super(GCEConnection, self).request(*args, **kwargs)
+
+        # If gce_params has been set, then update the pageToken with the
+        # nextPageToken so it can be used in the next request.
+        if self.gce_params:
+            if 'nextPageToken' in response.object:
+                self.gce_params['pageToken'] = response.object['nextPageToken']
+            elif 'pageToken' in self.gce_params:
+                del self.gce_params['pageToken']
+            self.gce_params = None
+
+        return response
+
+
+class GCEList(object):
+    """
+    An Iterator that wraps list functions to provide additional features.
+
+    GCE enforces a limit on the number of objects returned by a list operation,
+    so users with more than 500 objects of a particular type will need to use
+    filter(), page() or both.
+
+    >>> l=GCEList(driver, driver.ex_list_urlmaps)
+    >>> for sublist in l.filter('name eq ...-map').page(1):
+    ...   sublist
+    ...
+    [<GCEUrlMap id="..." name="cli-map">]
+    [<GCEUrlMap id="..." name="web-map">]
+
+    One can create a GCEList manually, but it's slightly easier to use the
+    ex_list() method of :class:`GCENodeDriver`.
+    """
+
+    def __init__(self, driver, list_fn, **kwargs):
+        """
+        :param  driver: An initialized :class:``GCENodeDriver``
+        :type   driver: :class:``GCENodeDriver``
+
+        :param  list_fn: A bound list method from :class:`GCENodeDriver`.
+        :type   list_fn: ``instancemethod``
+        """
+        self.driver = driver
+        self.list_fn = list_fn
+        self.kwargs = kwargs
+        self.params = {}
+
+    def __iter__(self):
+        list_fn = self.list_fn
+        more_results = True
+        while more_results:
+            self.driver.connection.gce_params = self.params
+            yield list_fn(**self.kwargs)
+            more_results = 'pageToken' in self.params
+
+    def __repr__(self):
+        return '<GCEList list="%s" params="%s">' % (
+            self.list_fn.__name__, repr(self.params))
+
+    def filter(self, expression):
+        """
+        Filter results of a list operation.
+
+        GCE supports server-side filtering of resources returned by a list
+        operation. Syntax of the filter expression is fully descripted in the
+        GCE API reference doc, but in brief it is::
+
+            FIELD_NAME COMPARISON_STRING LITERAL_STRING
+
+        where FIELD_NAME is the resource's property name, COMPARISON_STRING is
+        'eq' or 'ne', and LITERAL_STRING is a regular expression in RE2 syntax.
+
+        >>> for sublist in l.filter('name eq ...-map'):
+        ...   sublist
+        ...
+        [<GCEUrlMap id="..." name="cli-map">, \
+                <GCEUrlMap id="..." name="web-map">]
+
+        API reference: https://cloud.google.com/compute/docs/reference/latest/
+        RE2 syntax: https://github.com/google/re2/blob/master/doc/syntax.txt
+
+        :param  expression: Filter expression described above.
+        :type   expression: ``str``
+
+        :return: This :class:`GCEList` instance
+        :rtype:  :class:`GCEList`
+        """
+        self.params['filter'] = expression
+        return self
+
+    def page(self, max_results=500):
+        """
+        Limit the number of results by each iteration.
+
+        This implements the paging functionality of the GCE list methods and
+        returns this GCEList instance so that results can be chained:
+
+        >>> for sublist in GCEList(driver, driver.ex_list_urlmaps).page(2):
+        ...   sublist
+        ...
+        [<GCEUrlMap id="..." name="cli-map">, \
+                <GCEUrlMap id="..." name="lc-map">]
+        [<GCEUrlMap id="..." name="web-map">]
+
+        :keyword  max_results: Maximum number of results to return per
+                               iteration. Defaults to the GCE default of 500.
+        :type     max_results: ``int``
+
+        :return: This :class:`GCEList` instance
+        :rtype:  :class:`GCEList`
+        """
+        self.params['maxResults'] = max_results
+        return self
 
 
 class GCELicense(UuidMixin):
@@ -1032,6 +1185,25 @@ class GCENodeDriver(NodeDriver):
         response = self.connection.request(request, method='GET').object
         return response['contents']
 
+    def ex_list(self, list_fn, **kwargs):
+        """
+        Wrap a list method in a :class:`GCEList` iterator.
+
+        >>> for sublist in driver.ex_list(driver.ex_list_urlmaps).page(1):
+        ...   sublist
+        ...
+        [<GCEUrlMap id="..." name="cli-map">]
+        [<GCEUrlMap id="..." name="lc-map">]
+        [<GCEUrlMap id="..." name="web-map">]
+
+        :param  list_fn: A bound list method from :class:`GCENodeDriver`.
+        :type   list_fn: ``instancemethod``
+
+        :return: An iterator that returns sublists from list_fn.
+        :rtype: :class:`GCEList`
+        """
+        return GCEList(driver=self, list_fn=list_fn, **kwargs)
+
     def ex_list_disktypes(self, zone=None):
         """
         Return a list of DiskTypes for a zone or all.
@@ -1128,10 +1300,13 @@ class GCENodeDriver(NodeDriver):
         project = self.ex_get_project()
         current_metadata = project.extra['commonInstanceMetadata']
         fingerprint = current_metadata['fingerprint']
+        md_items = []
+        if 'items' in current_metadata:
+            md_items = current_metadata['items']
 
         # grab copy of current 'sshKeys' in case we want to retain them
         current_keys = ""
-        for md in current_metadata['items']:
+        for md in md_items:
             if md['key'] == 'sshKeys':
                 current_keys = md['value']
 
@@ -2081,7 +2256,7 @@ class GCENodeDriver(NodeDriver):
         :type     ex_metadata: ``dict`` or ``None``
 
         :keyword  ex_boot_disk: The boot disk to attach to the instance.
-        :type     ex_boot_disk: :class:`StorageVolume` or ``str``
+        :type     ex_boot_disk: :class:`StorageVolume` or ``str`` or ``None``
 
         :keyword  use_existing_disk: If True and if an existing disk with the
                                      same name/location is found, use that
@@ -2093,7 +2268,7 @@ class GCENodeDriver(NodeDriver):
                                used.  If 'None', then no external address will
                                be used.  To use an existing static IP address,
                                a GCEAddress object should be passed in.
-        :type     external_ip: :class:`GCEAddress` or ``str`` or None
+        :type     external_ip: :class:`GCEAddress` or ``str`` or ``None``
 
         :keyword  ex_disk_type: Specify a pd-standard (default) disk or pd-ssd
                                 for an SSD disk.
@@ -2184,6 +2359,8 @@ n
             image = self.ex_get_image(image)
         if not hasattr(ex_disk_type, 'name'):
             ex_disk_type = self.ex_get_disktype(ex_disk_type, zone=location)
+        if ex_boot_disk and not hasattr(ex_boot_disk, 'name'):
+            ex_boot_disk = self.ex_get_disk(ex_boot_disk, zone=location)
 
         # Use disks[].initializeParams to auto-create the boot disk
         if not ex_disks_gce_struct and not ex_boot_disk:
@@ -4389,7 +4566,9 @@ n
                                (default), a new non-static address will be
                                used.  If 'None', then no external address will
                                be used.  To use an existing static IP address,
-                               a GCEAddress object should be passed in.
+                               a GCEAddress object should be passed in. This
+                               param will be ignored if also using the
+                               ex_nic_gce_struct param.
         :type     external_ip: :class:`GCEAddress` or ``str`` or None
 
         :keyword  ex_disk_type: Specify a pd-standard (default) disk or pd-ssd
@@ -4427,7 +4606,7 @@ n
                                        is made to ensure proper formatting of
                                        the disks[] structure. Using this
                                        structure obviates the need of using
-                                       other disk params like 'ex_boot_disk',
+                                       other disk params like 'boot_disk',
                                        etc. See the GCE docs for specific
                                        details.
         :type     ex_disks_gce_struct: ``list`` or ``None``
@@ -4526,9 +4705,21 @@ n
         if ex_disks_gce_struct:
             node_data['disks'] = ex_disks_gce_struct
 
-        if network and ex_nic_gce_struct:
-            raise ValueError("Cannot specify both 'network' and "
-                             "'ex_nic_gce_struct'. Use one or the other.")
+        if ex_nic_gce_struct is not None:
+            if hasattr(external_ip, 'address'):
+                raise ValueError("Cannot specify both a static IP address "
+                                 "and 'ex_nic_gce_struct'. Use one or the "
+                                 "other.")
+            if hasattr(network, 'name'):
+                if network.name == 'default':
+                    # assume this is just the default value from create_node()
+                    # and since the user specified ex_nic_gce_struct, the
+                    # struct should take precedence
+                    network = None
+                else:
+                    raise ValueError("Cannot specify both 'network' and "
+                                     "'ex_nic_gce_struct'. Use one or the "
+                                     "other.")
 
         if network:
             ni = [{'kind': 'compute#instanceNetworkInterface',
@@ -4654,7 +4845,7 @@ n
         request, node_data = self._create_node_req(
             status['name'], node_attrs['size'], node_attrs['image'],
             node_attrs['location'], node_attrs['network'], node_attrs['tags'],
-            node_attrs['metadata'], boot_disk=status['disk'],
+            node_attrs['metadata'],
             external_ip=node_attrs['external_ip'],
             ex_service_accounts=node_attrs['ex_service_accounts'],
             description=node_attrs['description'],
@@ -5071,6 +5262,7 @@ n
         extra['canIpForward'] = node.get('canIpForward')
         extra['serviceAccounts'] = node.get('serviceAccounts', [])
         extra['scheduling'] = node.get('scheduling', {})
+        extra['boot_disk'] = None
 
         for disk in extra['disks']:
             if disk.get('boot') and disk.get('type') == 'PERSISTENT':
@@ -5090,10 +5282,17 @@ n
 
         # For the node attributes, use just machine and image names, not full
         # paths.  Full paths are available in the "extra" dict.
+        image = None
         if extra['image']:
             image = self._get_components_from_path(extra['image'])['name']
         else:
-            image = None
+            if extra['boot_disk'] and \
+                    hasattr(extra['boot_disk'], 'extra') and \
+                    'sourceImage' in extra['boot_disk'].extra and \
+                    extra['boot_disk'].extra['sourceImage'] is not None:
+                src_image = extra['boot_disk'].extra['sourceImage']
+                image = self._get_components_from_path(src_image)['name']
+            extra['image'] = image
         size = self._get_components_from_path(node['machineType'])['name']
 
         return Node(id=node['id'], name=node['name'],
@@ -5231,6 +5430,15 @@ n
         extra['status'] = volume.get('status')
         extra['creationTimestamp'] = volume.get('creationTimestamp')
         extra['description'] = volume.get('description')
+        extra['sourceImage'] = volume.get('sourceImage')
+        extra['sourceImageId'] = volume.get('sourceImageId')
+        extra['sourceSnapshot'] = volume.get('sourceSnapshot')
+        extra['sourceSnapshotId'] = volume.get('sourceSnapshotId')
+        extra['options'] = volume.get('options')
+        if 'licenses' in volume:
+            lic_objs = self._licenses_from_urls(licenses=volume['licenses'])
+            extra['licenses'] = lic_objs
+
         extra['type'] = volume.get('type', 'pd-standard').split('/')[-1]
 
         return StorageVolume(id=volume['id'], name=volume['name'],