+++ libcloud/compute/drivers/openstack.py
@@ -1195,6 +1195,11 @@ class OpenStack_1_1_NodeDriver(OpenStackNodeDriver):
                                  https://help.ubuntu.com/community/CloudInit
         :type       ex_userdata: ``str``
 
+        :keyword    ex_config_drive: Enable config drive
+                                     see
+                                     http://docs.openstack.org/grizzly/openstack-compute/admin/content/config-drive.html
+        :type       ex_config_drive: ``bool``
+
         :keyword    ex_security_groups: List of security groups to assign to
                                         the node
         :type       ex_security_groups: ``list`` of
@@ -1304,6 +1309,9 @@ class OpenStack_1_1_NodeDriver(OpenStackNodeDriver):
         if 'ex_userdata' in kwargs:
             server_params['user_data'] = base64.b64encode(
                 b(kwargs['ex_userdata'])).decode('ascii')
+
+        if 'ex_config_drive' in kwargs:
+            server_params['config_drive'] = kwargs['ex_config_drive']
 
         if 'ex_disk_config' in kwargs:
             server_params['OS-DCF:diskConfig'] = kwargs['ex_disk_config']