+++ CHANGES.rst
@@ -25,6 +25,10 @@ General
 Compute
 ~~~~~~~
 
+- Google Compute now supports paginated lists including filtering.
+  (GITHUB-491)
+  [Lee Verberne]
+
 - OpenStackNodeSize objects now support optional, additional fields that are
   supported in OpenStack 2.1: `ephemeral_disk`, `swap`, `extra`.
   (GITHUB-488, LIBCLOUD-682)
@@ -154,7 +158,7 @@ Compute
   [Avi Nanhkoesingh]
 
 - Add new driver for the Dimension Data provider based on the OpSource driver.
-  (LIBCLOUD-698, GITHUB-507)
+  (LIBCLOUD-698, GITHUB-507, LIBCLOUD-700, GITHUB-513)
   [Anthony Shaw]
 
 - Add "virtualmachine_id" attribute to the ``CloudStackAddress`` class in the
@@ -166,6 +170,25 @@ Compute
   ``ex_list_port_forwarding_rules`` in the CloudStack driver.
   (LIBCLOUD-678, GITHUB-484)
   [Atsushi Sasaki]
+
+- Fix an issue with ``list_nodes`` in the CloudSigma driver throwing an
+  exception if a node in the list had a static IP.
+  (LIBCLOUD-707, GITHUB-514)
+  [Chris O'Brien]
+
+- Don't throw an exception if a limit for a particular CloudStack resource is
+  "Unlimited" and not a number.
+  (GITHUB-512)
+  [Syed Mushtaq Ahmed]
+
+- Allow user to pass ``ex_config_drive`` argument to the ``create_node`` method
+  in the OpenStack driver.
+  (LIBCLOUD-356, GITHUB-330)
+  [Ryan Parrish]
+
+- Add new driver for Cloudwatt (https://www.cloudwatt.com/en/) provider.
+  (GITHUB-338)
+  [ZuluPro]
 
 Storage
 ~~~~~~~